+++ e2e/wdio/core/pages/multi-input.po.ts
@@ -3,7 +3,6 @@ import { CoreBaseComponentPo } from './core-base-component.po';
 
 export class MultiInputPo extends CoreBaseComponentPo {
     private url = '/multi-input';
-    root = '#page-content';
 
     activeDropdownButtons = 'button[title="value-help"]:not(:disabled)';
     activeInputs = '.fd-multi-input-field input:not(:disabled)';
@@ -31,6 +30,7 @@ export class MultiInputPo extends CoreBaseComponentPo {
     dialogCheckbox = '.fd-dialog fd-checkbox';
     selectAllItemsBtn = '.fd-button[glyph="multiselect-all"]';
     dialogListItem = '.fd-dialog .fd-list__item';
+    compactInput = '.fd-input--compact';
 
     open(): void {
         super.open(this.url);