+++ libs/core/schematics/ng-add/schema.ts
@@ -1,4 +1,5 @@
 export interface Schema {
     project: string;
     animations: boolean;
+    styleFonts: boolean;
 }