+++ libs/platform/schematics/utils/package-utils.ts
@@ -37,6 +37,17 @@ export function hasPackage(tree: Tree, name: string): boolean | null {
     return packageJson.dependencies && packageJson.dependencies[name];
 }
 
+// Check if a package exists in the package.json
+export function hasDevPackage(tree: Tree, name: string): boolean | null {
+    if (!tree.exists('package.json')) {
+        return null;
+    }
+
+    const packageJson = JSON.parse(tree.read('package.json')!.toString('utf-8'));
+
+    return packageJson.devDependencies && packageJson.devDependencies[name];
+}
+
 // Returns the source path for the application
 export async function getSourceTreePath(host: Tree, options: any): Promise<string> {
     const project = await getWorkspaceProject(host, options);
@@ -90,16 +101,4 @@ export async function getWorkspaceProject(host: Tree, options: any): Promise<wor
         throw new SchematicsException(`Invalid project name: ${options.project}`);
     }
     return project;
-}
-
-// Returns the default project for the application
-export async function getDefaultProject(host: Tree, options: any): Promise<any> {
-    const workspaceHost = createHost(host);
-    const { workspace } = await workspaces.readWorkspace('/', workspaceHost);
-
-    if (!options.project) {
-        options.project = workspace.extensions.defaultProject;
-    }
-
-    return options.project;
 }